refactoring: splitting apart SyncthingVirtualFolderAccessI and its impl
